+++ final_pjt_back/outerapi/models.py
@@ -22,7 +22,7 @@ class DepositProducts(models.Model):
     dcls_strt_day = models.CharField(max_length=50)
 
 class DepositOptions(models.Model):
-    product = models.ForeignKey(DepositProducts, on_delete=models.CASCADE,related_name='options')
+    product = models.ForeignKey(DepositProducts, on_delete=models.CASCADE,related_name='deposit_options')
     fin_prdt_cd = models.TextField() 
     intr_rate_type_nm = models.CharField(max_length=100) 
     intr_rate = models.FloatField() 
@@ -43,7 +43,7 @@ class SavingProducts(models.Model):
     dcls_strt_day = models.CharField(max_length=50)
 
 class SavingOptions(models.Model):
-    product = models.ForeignKey(DepositProducts, on_delete=models.CASCADE,related_name='options')
+    product = models.ForeignKey(SavingProducts, on_delete=models.CASCADE,related_name='saving_options')
     fin_prdt_cd = models.TextField() 
     intr_rate_type_nm = models.CharField(max_length=100) 
     intr_rate = models.FloatField()
